Remove dataManager and move client_test.go from gquic/ to self/
