+++ integrationtests/gquic/random_rtt_test.go
@@ -51,10 +51,6 @@ var _ = Describe("Random Duration Generator", func() {
 })
 
 var _ = Describe("Random RTT", func() {
-	BeforeEach(func() {
-		dataMan.GenerateData(dataLen)
-	})
-
 	var proxy *quicproxy.QuicProxy
 
 	runRTTTest := func(minRtt, maxRtt time.Duration, version protocol.VersionNumber) {
@@ -73,14 +69,14 @@ var _ = Describe("Random RTT", func() {
 			"--quic-version="+strconv.Itoa(int(version)),
 			"--host=127.0.0.1",
 			"--port="+strconv.Itoa(proxy.LocalPort()),
-			"https://quic.clemente.io/data",
+			"https://quic.clemente.io/prdata",
 		)
 
 		session, err := Start(command, nil, GinkgoWriter)
 		Expect(err).NotTo(HaveOccurred())
 		defer session.Kill()
 		Eventually(session, 20).Should(Exit(0))
-		Expect(bytes.Contains(session.Out.Contents(), dataMan.GetData())).To(BeTrue())
+		Expect(bytes.Contains(session.Out.Contents(), testserver.PRData)).To(BeTrue())
 	}
 
 	AfterEach(func() {
